+++ suite_context_test.go
@@ -73,6 +73,15 @@ func InitializeScenario(ctx *ScenarioContext) {
 	ctx.Step(`^(?:a )?passing step$`, func() error {
 		return nil
 	})
+	ctx.Given(`^(?:a )?given step$`, func() error {
+		return nil
+	})
+	ctx.When(`^(?:a )?when step$`, func() error {
+		return nil
+	})
+	ctx.Then(`^(?:a )?then step$`, func() error {
+		return nil
+	})
 
 	// Introduced to test formatter/cucumber.feature
 	ctx.Step(`^the rendered json will be as follows:$`, tc.theRenderJSONWillBe)
@@ -91,8 +100,16 @@ func InitializeScenario(ctx *ScenarioContext) {
 		return Steps{"passing step", "undefined step", "passing step"}
 	})
 
+	ctx.Then(`^(?:a |an )?undefined multistep using 'then' function$`, func() Steps {
+		return Steps{"given step", "undefined step", "then step"}
+	})
+
 	ctx.Step(`^(?:a )?passing multistep$`, func() Steps {
 		return Steps{"passing step", "passing step", "passing step"}
+	})
+
+	ctx.Then(`^(?:a )?passing multistep using 'then' function$`, func() Steps {
+		return Steps{"given step", "when step", "then step"}
 	})
 
 	ctx.Step(`^(?:a )?failing nested multistep$`, func() Steps {